+++ fetools-app/src/components/ColorGradientSlider.jsx
